refactor: Group the related code in the same field
