--- tests/Feature/Appointments/RelationClinicAndDoctorTest.php
@@ -1,67 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Feature\Appointments;
-
-use Database\Factories\ClinicFactory;
-use Database\Factories\DoctorFactory;
-use Database\Factories\UserFactory;
-use Lightit\Appointments\App\Exceptions\RelationException;
-use Lightit\Appointments\Domain\Actions\UpsertAppointmentAction;
-use Lightit\Appointments\Domain\Actions\ValidateClinicDoctorRelation;
-use Lightit\Appointments\Domain\Actions\ValidateDoctorOverlapping;
-use Lightit\Appointments\Domain\Actions\ValidateUserOverlapping;
-use Lightit\Appointments\Domain\DataTransferObjects\AppointmentDto;
-
-describe('clinic-doctor relation validation', function (): void {
-    it('throw an exception if the doctor does not work at the selected clinic', function (): void {
-        $doctor = DoctorFactory::new()->createOne();
-        $user = UserFactory::new()->createOne();
-        $clinic = ClinicFactory::new()->createOne();
-
-        $dto = new AppointmentDto(
-            doctorId: $doctor->id,
-            clinicId: $clinic->id,
-            startTime: now()->toDateTimeString(),
-            endTime: now()->addHour()->toDateTimeString()
-        );
-
-        $action = new UpsertAppointmentAction(
-            new ValidateDoctorOverlapping(),
-            new ValidateUserOverlapping(),
-            new ValidateClinicDoctorRelation(),
-        );
-
-        try {
-            $action->execute($dto, $user);
-            $this->fail('Expecte RelationException was not thrown.');
-        } catch (RelationException $e) {
-            expect($e->getMessage())->tobe('This doctor does not work at this clinic');
-        }
-    });
-
-    it('does not throw RelationException when the Doctor belong to a Clinic', function (): void {
-        $doctor = DoctorFactory::new()->createOne();
-        $user = UserFactory::new()->createOne();
-        $clinic = ClinicFactory::new()->createOne();
-        $clinic->doctors()->attach($doctor);
-
-        $dto = new AppointmentDto(
-            doctorId: $doctor->id,
-            clinicId: $clinic->id,
-            startTime: now()->toDateTimeString(),
-            endTime: now()->addHour()->toDateTimeString()
-        );
-
-        $action = new UpsertAppointmentAction(
-            new ValidateDoctorOverlapping(),
-            new ValidateUserOverlapping(),
-            new ValidateClinicDoctorRelation(),
-        );
-
-        expect(fn (): \Lightit\Appointments\Domain\Models\Appointment => $action->execute($dto, $user))->not()->toThrow(
-            RelationException::class
-        );
-    });
-});
